New PMAP implementation for MIPS64.
- Add referenced bit emulation.  The hardware valid bit in the TLB entry
is used as a referenced bit and a software valid bit is added
(PTE_SV).  On a TLB exception the software valid bit is checked and the
hardware valid/referenced bit (PTE_VR) is set.  The hardware valid bit is
effectively used as a referenced bit.

- Add support for automatic promotion of 4KB page mappings to 2MB page
mappings (aka. "superpages").  Automatic promotion can be enabled by
setting the tunable "vm.pmap.pg_ps_enabled" to a non-zero value.  By
default, automatic promotion is disabled. (Further testing is needed
before it is enabled by default.)

On MIPS64 the 2MB superpages are actually an even and odd pair of 1M
pages that act as a single 2MB superpage mapping.  This allows the VM layer
to believe it is just a single 2M superpage and not have to deal with the
way the TLB is implemented on MIPS in a special way.  The 4K pages are
still mapped in the TLB as before (i.e. even and odd contiguous pages in
virtual memory share a single TLB entry but may mapped non-contiguous
physical memory.)

- PV chunk and list locking so many of the pmap_* functions are no longer
serialized by the pvh global lock just like in the AMD64 implementation.
Many of the other optimizations to the AMD64 pmap implementation are also
included as well.

- Add "Machine Check" exception handling and recovery.  If for some
reason conflicting TLB are added and the CPU generates a "Machine Check"
exception then the trap handler will detect and recover by flushing the
TLB and reseting the status register.  Also, report the MCHECK.

- Use large (16K sized pages) for the kernel thread stack.  To prevent a
kernel thread stack overrun a larger thread stack is used.  The larger
stack is created by using a larger sized page (16K instead of 4K).

The new pmap implementation for MIPS64 is enabled by adding "options
MIPS64_NEW_PMAP" to the kernel config file.  The larger kernel thread
stack is enabled by adding "options KSTACK_LARGE_PAGE".